ingest/requirements.txt
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
python-dotenv>=1.0.0
structlog>=24.1.0
backoff>=2.2.1